pyproject.toml
[project]
name = "fyp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.57.1",
    "codebleu>=0.7.0",
    "datasets>=3.6.0",
    "matplotlib>=3.10.3",
    "nltk>=3.9.1",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "python-levenshtein>=0.27.1",
    "requests>=2.25.0",
    "rouge-score>=0.1.2",
    "scikit-learn>=1.7.0",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "tree-sitter==0.23",
    "tree-sitter-python>=0.23.6",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

[tool.uv.sources]
codebleu = { git = "https://github.com/k4black/codebleu.git" }
